refactor: 예외 핸들러 통합
